requirements.txt
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
coverage==7.7.0
idna==3.10
iniconfig==2.1.0
name-formatter @ git+https://github.com/philhanna/name-formatter.git@0f5a6f459662ed427bffd98d29473ceb7a66dbd2
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
pytest-cov==6.0.0
PyYAML==6.0.2
requests==2.32.3
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
